alpha implement "ignore conflict" and "force overlap" in the "add session" feature

+++ teacherview.controller.php
@@ -17,7 +17,7 @@ function scheduler_action_doaddsession($scheduler, $formdata) {
     global $DB, $output;
 
     $data = (object) $formdata;
-
+var_dump($data);
     $fordays = 0;
     if ($data->rangeend > 0){
         $fordays = ($data->rangeend - $data->rangestart) / DAYSECS;
@@ -74,7 +74,10 @@ function scheduler_action_doaddsession($scheduler, $formdata) {
             while ($slot->starttime <= $data->timeend - $slot->duration * 60) {
                 $conflicts = scheduler_get_conflicts($scheduler->id, $data->timestart, $data->timestart + $slot->duration * 60, $data->teacherid, 0, SCHEDULER_ALL, false);
                 if ($conflicts) {
-                    if (!$data->forcewhenoverlap) {
+                    //if (!$data->forcewhenoverlap) {
+                    var_dump(intval($data->conflictoptions));
+                    var_dump(SLOT_CONFLICT_PROHIBIT);
+                    if (intval($data->conflictoptions)==SLOT_CONFLICT_PROHIBIT) {
                         print_string('conflictingslots', 'scheduler');
                         echo '<ul>';
                         foreach ($conflicts as $aconflict) {
@@ -88,14 +91,17 @@ function scheduler_action_doaddsession($scheduler, $formdata) {
                             echo html_writer::tag('li', $msg);
                         }
                         echo '</ul><br/>';
-                    } else { // we force, so delete all conflicting before inserting
+                    } 
+                    if (intval($data->conflictoptions)==SLOT_CONFLICT_FORCE_OVERLAP) { 
+                        // we force, so delete all conflicting before inserting
                         foreach ($conflicts as $conflict) {
                         	$cslot = $scheduler->get_slot($conflict->id);
                             $cslot->delete();
                         }
                     }
                 }
-                if (!$conflicts || $data->forcewhenoverlap) {
+                //if (!$conflicts || $data->forcewhenoverlap) {
+                if (!$conflicts || intval($data->conflictoptions)!=SLOT_CONFLICT_PROHIBIT) {
                     $DB->insert_record('scheduler_slots', $slot, false, true);
                     $countslots++;
                 }
